pyproject.toml
[tool.poetry]
name = "theskepticjudge"
version = "0.1.0"
description = "A project for skeptical agent evaluation"
authors = ["Varshini Balaji"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
openai = "^1.82.1"
galileo = {extras = ["openai"], version = "^0.10.0"}
pyyaml = "^6.0.2"
python-dotenv = "^1.1.0"
ipykernel = "^6.29.5"
pymongo = {version = "3.11", extras = ["srv"]}
pdfplumber = "^0.11.6"
requests = "^2.32.3"
fastapi = "^0.115.12"
uvicorn = "^0.34.2"
pydantic = "^2.11.5"
httpx = "^0.28.1"
google-generativeai = "^0.8.5"
numpy = "^2.2.6"
sentence-transformers = "^4.1.0"
einops = "^0.8.1"
tqdm = "^4.67.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
